object_dectection/requirements.txt
streamlit
pillow 
numpy
opencv-python